More refactoring on splitHX
